baykarproject/iha_kiralama_projesi/requirements.txt
Django
djangorestframework
django-bootstrap-datepicker-plus
django-filter
psycopg2
